Guides to finish adding hotspots more clearly
I've been spending a lot of time specifying clients that using this plugin means clicking the dull grey button "finish adding hotspots" and not the bright blue button "update".

This commit tries to improve the user experience so that the users are more clearly guided to click to finish button.

Update class-meauh-attachment.php